third_party/WebKit/LayoutTests/fast/dom/custom/lifecycle-created-createElement-reentrancy.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script>
description("This test ensures that the lifecycle callback of a parser-made element is visible in following script block.")

window.callbacksCalled = [];

function fooCreatedFunction() {
  shouldBe("window.callbacksCalled", "[]");
  window.callbacksCalled.push(this.tagName);
  this.innerHTML = "<x-bar></x-bar>";
  shouldBe("window.callbacksCalled", "['X-FOO', 'X-BAR']");
}

function barCreatedFunction() {
  shouldBe("window.callbacksCalled", "['X-FOO']");
  window.callbacksCalled.push(this.tagName);
}

document.registerElement("x-foo", { prototype: Object.create(HTMLElement.prototype, { createdCallback: { value: fooCreatedFunction } }) });
document.registerElement("x-bar", { prototype: Object.create(HTMLElement.prototype, { createdCallback: { value: barCreatedFunction } }) });
</script>
</head>
<body>
<script>
document.createElement("x-foo");
shouldBe("window.callbacksCalled", "['X-FOO', 'X-BAR']");
</script>
</body>
</html>
